LayoutTests/media/event-attributes-expected.txt

EVENT(loadstart)
EVENT(durationchange)
EVENT(loadedmetadata)
EVENT(loadeddata)
EVENT(canplay)
EVENT(canplaythrough)
EVENT(load)

RUN(video.play())
EVENT(play)
EVENT(playing)

RUN(video.playbackRate = 2)
EVENT(ratechange)

RUN(video.volume = 0.5)
EVENT(volumechange)

RUN(video.pause())
EVENT(timeupdate)
EVENT(pause)

RUN(video.currentTime = 5.6)
EVENT(timeupdate)
EVENT(seeked)

RUN(video.play())
EVENT(ratechange)
EVENT(play)
EVENT(playing)
EVENT(timeupdate)
EVENT(timeupdate)
EVENT(timeupdate)
EVENT(ended)

RUN(video.src = 'content/garbage.mp4')
RUN(video.load())
EVENT(emptied)
EVENT(loadstart)
EVENT(error)
END OF TEST
